+++ src/helpers/authHelper.js
@@ -1,4 +1,3 @@
-// import englishWiseApi from '../services/englishWiseApi';
 import { userSession } from '../services/englishWiseApi';
 
 const USER_TOKEN = 'userToken';
